API

Modules

MonotonicSplines.MonotonicSplines

Types and constants

MonotonicSplines.InvRQSpline

MonotonicSplines.RQSpline

Functions and macros

MonotonicSplines._sort_dimensions

MonotonicSplines.eval_forward_rqs_params

MonotonicSplines.eval_forward_rqs_params_with_grad

MonotonicSplines.get_params

MonotonicSplines.rqs_backward

MonotonicSplines.rqs_forward

MonotonicSplines.rqs_forward_kernel!

MonotonicSplines.rqs_forward_pullback

MonotonicSplines.rqs_forward_pullback_kernel!

MonotonicSplines.spline_backward

MonotonicSplines.spline_forward

Documentation

MonotonicSplines.MonotonicSplines — Module

MonotonicSplines

High-performance monotonic splines in Julia.

MonotonicSplines.InvRQSpline — Type

InvRQSpline(widths::AbstractArray{<:Real}, heights::AbstractArray{<:Real}, derivatives::AbstractArray{<:Real})

Object holding the parameters to characterize several inverse rational quadratic spline functions analogous to InvRQSpline.

The same parameters are used to characterize the forward and inverse spline functions, the struct used to store them decides the equation they are evaluated in.

MonotonicSplines.RQSpline — Type

RQSpline(widths::AbstractArray{<:Real}, heights::AbstractArray{<:Real}, derivatives::AbstractArray{<:Real})

Object holding the parameters to characterize several rational quadratic spline functions after the scheme first defined by Gregory and Delbourgo in https://doi.org/10.1093/imanum/2.2.123. RQSpline holds the parameters to characterize n_dims x n_samples spline functions to transform each component of n_samples samples.

widths, heights, and derivatives are K+1 x n_dims x n_samples arrays, with the parameters to characterize a single spline function with K segments in the first dimension. Along the second dimension, the spline functions for a single sample are stored, and along the third dimension the sets splines for differen samples.

MonotonicSplines._sort_dimensions — Method

sort_dimensions(y₁::AbstractArray, y₂::AbstractArray, mask::AbstractVector)

Output an array identical to y₂, but with the rows specified by mask overwritten with the corresponding rows in y₁.

MonotonicSplines.eval_forward_rqs_params — Method

eval_forward_rqs_params(wₖ::Real, wₖ₊₁::Real, hₖ::Real, hₖ₊₁::Real, dₖ::Real, dₖ₊₁::Real, x::Real)

Apply a rational quadratic spline segment to x, and calculate the logarithm of the absolute value of this function's jacobian.

MonotonicSplines.eval_forward_rqs_params_with_grad — Method

eval_forward_rqs_params_with_grad(wₖ::Real, wₖ₊₁::Real, hₖ::Real, hₖ₊₁::Real, dₖ::Real, dₖ₊₁::Real, x::Real)

Apply a rational quadratic spline segment to x, and calculate the logarithm of the absolute value of this function's jacobian. And calculate the gradient of that function depending on the spline parameters.

MonotonicSplines.get_params — Function

get_params(θ_raw::AbstractArray, n_dims_trafo::Integer, B::Real = 5.)

Process the raw output parameters of a neural net to be usable as the widths, heights and derivatives to characterize a set of rational quadratic spline functions. Outputs a tripel w,h,d containing the widths, heights, and derivative parameters, either on the CPU or GPU, depending on the backend of the input θ_raw.

The θ_raw should be a matrix, with the columns being the raw parameters for a sample. n_dims_trafo is the number of spline functions for which parameters are supposed to be produced. The output parameters are stored in an K+1 x n_spline_functions_per_sample x n_samples array.

MonotonicSplines.rqs_backward — Method

rqs_backward(x::AbstractArray{<:Real}, w::AbstractArray{<:Real}, h::AbstractArray{<:Real}, d::AbstractArray{<:Real}, w_logJac::AbstractArray{<:Real}, h_logJac::AbstractArray{<:Real}, d_logJac::AbstractArray{<:Real})

Apply the inverse rational quadratic spline functions characterized by w, h, and d to x. The spline function characterized by the parameters in the [:,i,j] entries in trafo is applied to the [i,j]-th element of x.

Return the transformed values in a matrix y of the same shape as x, and return a 1 x size(x,2) -matrix holding the sums of the values of the logarithm of the absolute values of the determinant of the jacobians of the spline functions applied to a column of x.

The function executes in a kernel, on the same backend as x is stored (CPU or GPU), the output will also be returned on the same backend.

MonotonicSplines.rqs_forward — Method

rqs_forward(x::AbstractArray{<:Real}, w::AbstractArray{<:Real}, h::AbstractArray{<:Real}, d::AbstractArray{<:Real})

Apply the rational quadratic spline functions characterized by w, h, and d to x. The spline function characterized by the parameters in the [:,i,j] entries in trafo is applied to the [i,j]-th element of x.

Return the transformed values in a matrix y of the same shape as x, and return a 1 x size(x,2) -matrix holding the sums of the values of the logarithm of the absolute values of the determinant of the jacobians of the spline functions applied to a column of x.

The function executes in a kernel, on the same backend as x is stored (CPU or GPU), the output will also be returned on the same backend.

MonotonicSplines.rqs_forward_kernel! — Method

rqs_forward_kernel!(x::AbstractArray, y::AbstractArray, logJac::AbstractArray, w::AbstractArray, h::AbstractArray, d::AbstractArray)

Apply the rational quadratic spline functions characterized by w, h, and d to x. The spline function characterized by the parameters in the [:,i,j] entries in trafo is applied to the [i,j]-th element of x.

The transformed values are stored in y the sums of the values of the logarithm of the absolute values of the determinant of the jacobians of the spline functions applied to a column of x are stored in the corresponding column of logJac.

To find the bin k in which the respective x value for a spline falls in, a the corresponding column of w is searched.j

MonotonicSplines.rqs_forward_pullback — Method

rqs_forward_pullback(x::AbstractArray{<:Real}, w::AbstractArray{<:Real}, h::AbstractArray{<:Real}, d::AbstractArray{<:Real}, tangent_1::AbstractArray, tangent_2::AbstractArray)

Return the gradients of the spline functions characterized by w, h, and d, evaluated at the values in x. The output will be on the same backend as x and w, h, and d (CPU or GPU).

MonotonicSplines.rqs_forward_pullback_kernel! — Method

rqs_forward_pullback_kernel!(
    x::AbstractArray,
    y::AbstractArray,
    logJac::AbstractArray,
    w::AbstractArray,
    h::AbstractArray,
    d::AbstractArray,
    ∂y∂w_tangent::AbstractArray,
    ∂y∂h_tangent::AbstractArray,
    ∂y∂d_tangent::AbstractArray,
    ∂LogJac∂w_tangent::AbstractArray,
    ∂LogJac∂h_tangent::AbstractArray,
    ∂LogJac∂d_tangent::AbstractArray,
    tangent_1::AbstractArray,
    tangent_2::AbstractArray,
)

Return the gradients of the rational quadratic spline functions characterized by w, h, and d, evaluated at the values in x and of logJac. The output will be on the same backend as x and w, h, and d (CPU or GPU).

MonotonicSplines.spline_backward — Method

spline_backward(trafo::InvRQSpline, x::AbstractMatrix{<:Real})

Apply the inverse rational quadratic spline functions characterized by the parameters stored in trafo to the matrix x. The rational quadratic spline function characterized by the parameters in the [:,i,j] entries in trafo is applied to the [i,j]-th element of x.

MonotonicSplines.spline_forward — Method

spline_forward(trafo::RQSpline, x::AbstractMatrix{<:Real})

Apply the rational quadratic spline functions characterized by the parameters stored in trafo to the matrix x. The spline function characterized by the parameters in the [:,i,j] entries in trafo is applied to the [i,j]-th element of x.