API

Modules

MonotonicSplines.MonotonicSplines

Types and constants

MonotonicSplines.InvRQSpline

MonotonicSplines.RQSpline

Functions and macros

MonotonicSplines._sort_dimensions

MonotonicSplines.eval_backward_rqs_params

MonotonicSplines.eval_backward_rqs_params_with_grad

MonotonicSplines.eval_forward_rqs_params

MonotonicSplines.eval_forward_rqs_params_with_grad

MonotonicSplines.get_params

MonotonicSplines.rqs_backward

MonotonicSplines.rqs_backward_kernel!

MonotonicSplines.rqs_forward

MonotonicSplines.rqs_forward_kernel!

MonotonicSplines.rqs_pullback

MonotonicSplines.rqs_pullback_kernel!

MonotonicSplines.spline_backward

MonotonicSplines.spline_forward

Documentation

MonotonicSplines.MonotonicSplines — Module

MonotonicSplines

High-performance monotonic splines in Julia.

MonotonicSplines.InvRQSpline — Type

InvRQSpline(widths::AbstractArray{<:Real}, heights::AbstractArray{<:Real}, derivatives::AbstractArray{<:Real})

A struct that holds the parameters to characterize multiple inverse rational quadratic spline functions. These functions are used to transform each component of multiple samples, providing a high-performance solution for batch transformations.

The InvRQSpline struct is based on the scheme first defined by Gregory and Delbourgo (https://doi.org/10.1093/imanum/2.2.123) and is designed to handle n_dims x n_samples inverse spline functions.

Fields

widths: A K+1 x n_dims x n_samples array that holds the width parameters for each inverse spline function. The first dimension corresponds to the number of segments in a single spline function, the second dimension corresponds to the inverse spline functions for a single sample, and the third dimension corresponds to the sets of inverse splines for different samples.

heights: A K+1 x n_dims x n_samples array that holds the height parameters for each inverse spline function. The dimensions are organized in the same way as the widths array.

derivatives: A K+1 x n_dims x n_samples array that holds the derivative parameters for each inverse spline function. The dimensions are organized in the same way as the widths and heights arrays.

Usage

An instance of InvRQSpline can be used like a function to apply the characterized inverse spline transformations to the components of the samples.

Performance

The struct is designed to leverage parallel computing and automatic differentiation for high performance. It uses KernelAbstractions.jl for parallel execution on either a CPU or a GPU, and Zygote.jl for automatic differentiation, providing precise and efficient gradient computations.

Note

The same parameters are used to characterize both the forward (RQSpline) and inverse (InvRQSpline) spline functions. The struct used to store them determines the equation they are evaluated in.

MonotonicSplines.RQSpline — Type

RQSpline(widths::AbstractArray{<:Real}, heights::AbstractArray{<:Real}, derivatives::AbstractArray{<:Real})

A struct that holds the parameters to characterize multiple rational quadratic spline functions. These functions are used to transform each component of multiple samples, providing a high-performance solution for batch transformations.

The RQSpline struct is based on the scheme first defined by Gregory and Delbourgo (https://doi.org/10.1093/imanum/2.2.123) and is designed to handle n_dims x n_samples spline functions For a batch of n_samples n_dims-dimensional samples.

Fields

widths: A K+1 x n_dims x n_samples array that holds the width parameters for each spline function. The first dimension corresponds to the number of segments in a single spline function, the second dimension corresponds to the spline functions for a single sample, and the third dimension corresponds to the sets of splines for different samples.

heights: A K+1 x n_dims x n_samples array that holds the height parameters for each spline function. The dimensions are organized in the same way as the widths array.

derivatives: A K+1 x n_dims x n_samples array that holds the derivative parameters for each spline function. The dimensions are organized in the same way as the widths and heights arrays.

Usage

An instance of RQSpline can be used like a function to apply the characterized spline transformations to the components of the samples.

Performance

The struct is designed to leverage parallel computing and automatic differentiation for high performance. It uses KernelAbstractions.jl for parallel execution on either a CPU or a GPU, and different backends for automatic differentiation, providing precise and efficient gradient computations.

MonotonicSplines._sort_dimensions — Method

_sort_dimensions(y₁::AbstractArray, y₂::AbstractArray, mask::AbstractVector)

Create a new array by selectively replacing rows from y₂ with corresponding rows from y₁ based on a boolean mask, mask.

Arguments

y₁: An array from which rows are taken. It should have the same number of columns as y₂.

y₂: An array that serves as the base for the output. Rows specified by mask are replaced with corresponding rows from y₁.

mask: A boolean vector of the same length as the number of rows in y₁ and y₂. If the i-th element of mask is true, the i-th row of y₂ is replaced with the i-th row of y₁ in the output.

Returns

res: An array of the same shape as y₂, but with rows specified by mask replaced with corresponding rows from y₁.

MonotonicSplines.eval_backward_rqs_params — Method

eval_forward_rqs_params(wₖ::Real, wₖ₊₁::Real, hₖ::Real, hₖ₊₁::Real, dₖ::Real, dₖ₊₁::Real, x::Real)

Apply a rational quadratic spline segment to a number x, and calculate the logarithm of the absolute value of this function's derivative.

Arguments

wₖ, wₖ₊₁: The width parameters of the spline segment at the edges of the k-th interval.

hₖ, hₖ₊₁: The height parameters of the spline segment.

dₖ, dₖ₊₁: The derivative parameters of the spline segment.

x: The value at which the spline function is evaluated.

Returns

y: The transformed value after applying the rational quadratic spline segment to x.

logJac: The logarithm of the absolute value of the derivative of the segment at x.

MonotonicSplines.eval_backward_rqs_params_with_grad — Method

eval_backward_rqs_params_with_grad(wₖ::M0, wₖ₊₁::M0, 
                                   hₖ::M1, hₖ₊₁::M1, 
                                   dₖ::M2, dₖ₊₁::M2, 
                                   x::M3            ) where {M0<:Real,M1<:Real, M2<:Real, M3<:Real}

Apply an inverse rational quadratic spline segment to x, calculate the logarithm of the absolute value of the derivative ("LogJac") of the segment at x, and compute the gradient of that segment and the LogJac with respect to the spline parameters.

Arguments

wₖ, wₖ₊₁: The width parameters of the spline segment at the edges of the k-th interval.

hₖ, hₖ₊₁: The height parameters of the spline segment.

dₖ, dₖ₊₁: The derivative parameters of the spline segment.

x: The value at which the spline function is evaluated.

Returns

y: The transformed value after applying the inverse rational quadratic spline segment to x.

logJac: The logarithm of the absolute value of the derivative of the segment at x.

∂y∂w, ∂y∂h, ∂y∂d: The gradients of y with respect to the two width, height, and derivative parameters, respectively.

∂LogJac∂w, ∂LogJac∂h, ∂LogJac∂d: The gradients of LogJac with respect to the two width, height, and derivative parameters, respectively.

MonotonicSplines.eval_forward_rqs_params — Method

eval_forward_rqs_params(wₖ::Real, wₖ₊₁::Real, hₖ::Real, hₖ₊₁::Real, dₖ::Real, dₖ₊₁::Real, x::Real)

Apply a rational quadratic spline segment to a number x, and calculate the logarithm of the absolute value of this function's Jacobian.

Arguments

wₖ, wₖ₊₁: The width parameters of the spline segment at the edges of the k-th interval.

hₖ, hₖ₊₁: The height parameters of the spline segment.

dₖ, dₖ₊₁: The derivative parameters of the spline segment.

x: The value at which the spline function is evaluated.

Returns

y: The transformed value after applying the rational quadratic spline segment to x.

logJac: The logarithm of the absolute value of the derivative of the segment at x.

MonotonicSplines.eval_forward_rqs_params_with_grad — Method

eval_forward_rqs_params_with_grad(wₖ::M0, wₖ₊₁::M0, hₖ::M1, hₖ₊₁::M1, dₖ::M2, dₖ₊₁::M2, x::M3) where {M0<:Real,M1<:Real, M2<:Real, M3<:Real}

Apply a rational quadratic spline segment to x, calculate the logarithm of the absolute value of the derivative ("LogJac") of the segment at x, and compute the gradient of that segment and the LogJac with respect to the spline parameters.

Arguments

wₖ, wₖ₊₁: The width parameters of the spline segment at the edges of the k-th interval.

hₖ, hₖ₊₁: The height parameters of the spline segment.

dₖ, dₖ₊₁: The derivative parameters of the spline segment.

x: The value at which the spline function is evaluated.

Returns

y: The transformed value after applying the rational quadratic spline segment to x.

logJac: The logarithm of the absolute value of the derivative of the segment at x.

∂y∂w, ∂y∂h, ∂y∂d: The gradients of y with respect to the two width, height, and derivative parameters, respectively.

∂LogJac∂w, ∂LogJac∂h, ∂LogJac∂d: The gradients of logJac with respect to the two width, height, and derivative parameters, respectively.

MonotonicSplines.get_params — Function

get_params(θ_raw::AbstractArray, n_dims_trafo::Integer, B::Real = 5.)

Process the raw output parameters of a neural network to generate parameters for a set of rational quadratic spline functions.

Arguments

θ_raw: A matrix where each column represents the raw parameters for a sample.

n_dims_trafo: The number of spline functions for which parameters are to be produced.

B: A real number used to initialize the widths and heights of the spline functions. Default is 5.

Returns

A tuple w, h, d containing the widths, heights, and derivative parameters of the spline functions. The parameters are stored in a K+1 x n_spline_functions_per_sample x n_samples array.

MonotonicSplines.rqs_backward — Method

rqs_backward(x::AbstractArray{<:Real}, w::AbstractArray{<:Real}, h::AbstractArray{<:Real}, d::AbstractArray{<:Real})

Apply the inverse rational quadratic spline functions, characterized by the parameters w (widths), h (heights), and d (derivatives), to the array x.

One sample is stored in a column of the input x.

Arguments

x: An array of real numbers to which the inverse spline functions are applied. The [i,j]-th element of x is transformed by the inverse spline function characterized by the parameters in the [:,i,j] entries in w, h, and d.

w, h, d: Arrays that hold the width, height, and derivative parameters of the inverse spline functions, respectively. Each of these is a 3D array with dimensions K+1 x n_dims x n_samples.

Returns

Two objects are returned:

y: A matrix of the same shape as x that holds the transformed values.

logJac: A 1 x size(x,2) matrix that holds the sums of the values of the logarithm of the absolute values of the determinant of the Jacobians of the inverse spline functions applied to a column of x.

MonotonicSplines.rqs_backward_kernel! — Method

rqs_backward_kernel!(x::AbstractArray, y::AbstractArray, logJac::AbstractArray, w::AbstractArray, h::AbstractArray, d::AbstractArray)

This kernel function applies the inverse rational quadratic spline functions characterized by the parameters w, h, and d to x.

Arguments

x: An array of real numbers to which the inverse spline functions are applied.

w, h, d: Arrays that hold the width, height, and derivative parameters of the inverse spline functions, respectively.

y: An array where the transformed values will be stored.

logJac: An array where the sums of the values of the logarithm of the absolute values of the determinant of the Jacobians of the inverse spline

functions applied to a column of x will be stored.

MonotonicSplines.rqs_forward — Method

rqs_forward(x::AbstractArray{<:Real}, w::AbstractArray{<:Real}, h::AbstractArray{<:Real}, d::AbstractArray{<:Real})

Apply the rational quadratic spline functions, characterized by the parameters w (widths), h (heights), and d (derivatives), to the array x.

One sample is stored in a column of the input x.

Arguments

x: An array of real numbers to which the spline functions are applied. The [i,j]-th element of x is transformed by the spline function characterized by the parameters in the [:,i,j] entries in w, h, and d.

w, h, d: Arrays that hold the width, height, and derivative parameters of the spline functions, respectively. Each of these is a 3D array with dimensions K+1 x n_dims x n_samples.

Returns

Two objects are returned:

y: A matrix of the same shape as x that holds the transformed values.

logJac: A 1 x size(x,2) matrix that holds the sums of the values of the logarithm of the absolute values of the determinant of the Jacobians of the spline functions applied to a column of x.

MonotonicSplines.rqs_forward_kernel! — Method

rqs_forward_kernel!(x::AbstractArray, y::AbstractArray, logJac::AbstractArray, w::AbstractArray, h::AbstractArray, d::AbstractArray)

This kernel function applies the rational quadratic spline functions characterized by the parameters w, h, and d to x.

Arguments

x: An array of real numbers to which the spline functions are applied.

w, h, d: Arrays that hold the width, height, and derivative parameters of the spline functions, respectively.

y: An array where the transformed values are stored.

logJac: An array where the sums of the values of the logarithm of the absolute values of the determinant of the Jacobians of the spline

functions applied to a column of x are stored.

Description

The function works by applying the spline function characterized by the parameters in the [:,i,j] entries in w, h, and d to the [i,j]-th element of x. The transformed values are stored in y and the sums of the values of the logarithm of the absolute values of the determinant of the Jacobians of the spline functions applied to a column of x are stored in logJac.

Note

To find the bin k in which the respective x value for a spline falls in, the corresponding column of w is searched.

MonotonicSplines.rqs_pullback — Method

rqs_pullback(param_eval_function::Function, x::AbstractArray{<:Real}, w::AbstractArray{<:Real}, h::AbstractArray{<:Real}, d::AbstractArray{<:Real}, tangent_1::AbstractArray, tangent_2::AbstractArray)

Compute the gradients of the rational quadratic spline functions characterized by w, h, and d, evaluated at the values in x with respect to w, h, and d.

This function is designed to make the transformation using Rational Quadratic Splines in this package automatically differentiable. Whether the gradients of the forward or backward spline functions are calculated is determined by the param_eval_function argument.

Arguments

param_eval_function: The function used to evaluate a spline segment. Different functions are used for the forward and backward spline functions.

x: An array of real numbers at which the spline functions are evaluated.

w, h, d: Arrays that hold the width, height, and derivative parameters of the spline functions, respectively.

tangent_1, tangent_2: Arrays that hold the tangent vectors for the transformed output and the log abs det jacobians respectively.

Returns

Three values are returned:

∂y∂w + ∂LogJac∂w: An array with the same shape as w, with the [:,j,k]-th element holding the gradient of the [j,k]-th element of y with respect to the width parameters plus the gradient of the logarithm of the absolute value of the derivative of this [j,k]-th element of y with respect to the [j,k]-th element of x, with respect to the width parameters. For Example, the [i,j,k] element of this array is ∂yⱼₖ/∂wᵢⱼₖ + ∂(log(abs(∂yⱼₖ/∂xⱼₖ)))/∂wᵢⱼₖ.

∂y∂h + ∂LogJac∂h: An array with the same shape as h, holding the same gradients as described above, but with respect to the height parameters.

∂y∂d + ∂LogJac∂d: An array with the same shape as d, holding the same gradients as described above, but with respect to the derivative parameters.

Note

Since only one segment of each spline is evaluated for one element of x, the returned gradients are filled with zeros, except for the parameters of the segment that is evaluated for that element of x. For example, if the [j,k] -th element of x falls in the l-th bin of the interval mask, the [:,j,k] entries in ∂y∂w + ∂LogJac∂w are all zero, except the [l,j,k] and [l+1,j,k] elements, which hold the (generally non-zero) values ∂yⱼₖ/∂wₗⱼₖ + ∂(log(abs(∂yⱼₖ/∂xⱼₖ)))/∂wₗⱼₖ and ∂yⱼₖ/∂wₗ₊₁ⱼₖ + ∂(log(abs(∂yⱼₖ/∂xⱼₖ)))/∂wₗ₊₁ⱼₖ respectively.

The function executes in a kernel, on the same backend as x is stored (CPU or GPU), and the output is also returned on the same backend.

MonotonicSplines.rqs_pullback_kernel! — Method

rqs_pullback_kernel(
    param_eval_function::Function,
    x::AbstractArray,
    y::AbstractArray,
    logJac::AbstractArray,
    w::AbstractArray,
    h::AbstractArray,
    d::AbstractArray,
    ∂y∂w_tangent::AbstractArray,
    ∂y∂h_tangent::AbstractArray,
    ∂y∂d_tangent::AbstractArray,
    ∂LogJac∂w_tangent::AbstractArray,
    ∂LogJac∂h_tangent::AbstractArray,
    ∂LogJac∂d_tangent::AbstractArray,
    tangent_1::AbstractArray,
    tangent_2::AbstractArray,
)

This kernel function calculates the gradients of the rational quadratic spline functions characterized by w, h, and d, evaluated at the values in x and of logJac.

Arguments

param_eval_function The function used to evaluate a spline segment. Different functions are used for the forward and backward passes.

x: An array of real numbers to which the spline functions are applied.

w, h, d: Arrays that hold the width, height, and derivative parameters of the spline functions, respectively.

y: An array where the transformed values are stored.

logJac: An array where the sums of the values of the logarithm of the absolute values of the determinant of the Jacobians of the spline functions applied to a column of x are stored.

∂y∂w_tangent, ∂y∂h_tangent, ∂y∂d_tangent: Arrays that will contain the gradients of the spline functions with respect to w, h, and d, respectively.

∂LogJac∂w_tangent, ∂LogJac∂h_tangent, ∂LogJac∂d_tangent: Arrays that will contain the gradients of logJac with respect to w, h, and d, respectively.

tangent_1, tangent_2: Arrays that hold the tangent vectors for the forward pass.

For an explanation of the shape and contents of the gradient arrays, see the documentation of the rqs_pullback() function.

Note

This function is a kernel function and is used within the rqs_forward_pullback function to calculate the gradients of the spline functions and logJac. It is not intended to be called directly by the user.

MonotonicSplines.spline_backward — Method

spline_backward(trafo::InvRQSpline, x::AbstractMatrix{<:Real})

Apply the inverse rational quadratic spline functions, characterized by the parameters stored in trafo, to the matrix x.

This function provides a high-performance solution for batch transformations, applying multiple inverse spline functions to a matrix of samples simultaneously. One sample is stored in a column of the input matrix x.

Arguments

trafo: An instance of InvRQSpline that holds the parameters of the inverse spline functions. These parameters are widths, heights, and derivatives, each of which is a 3D array with dimensions K+1 x n_dims x n_samples.

x: A matrix of real numbers to which the inverse spline functions are applied. The [i,j]-th element of x is transformed by the inverse spline function characterized by the parameters in the [:,i,j] entries in the parameter arrays stored in trafo.

Returns

Two objects are returned:

y: A matrix of the same shape as x that holds the transformed values.

logJac: A 1 x size(x,2) matrix that holds the sums of the values of the logarithm of the absolute values of the determinant of the Jacobians of the spline functions applied to a column of x.

MonotonicSplines.spline_forward — Method

spline_forward(trafo::RQSpline, x::AbstractMatrix{<:Real})

Apply the rational quadratic spline functions, characterized by the parameters stored in trafo, to the matrix x.

This function provides a high-performance solution for batch transformations, applying multiple spline functions to a matrix of samples simultaneously. One sample is stored in a column of the input matrix x.

Arguments

trafo: An instance of RQSpline that holds the parameters of the spline functions. These parameters are widths, heights, and derivatives, each of which is a 3D array with dimensions K+1 x n_dims x n_samples.

x: A matrix of real numbers to which the spline functions are applied. The [i,j]-th element of x is transformed by the spline function characterized by the parameters in the [:,i,j] entries in the parameter arrays stored in trafo.

Returns

Two objects are returned:

y: A matrix of the same shape as x that holds the transformed values.

logJac: A 1 x size(x,2) matrix that holds the sums of the values of the logarithm of the absolute values of the determinant of the Jacobians of the spline functions applied to a column of x.